Fix test and build failures for Go 1.25 on Linux
- Fix non-constant fmt.Errorf format strings in pkg/ivideon and pkg/tuya
  (Go 1.25 vet strictness)
- Fix ffmpeg test expectations: H265 software codec uses -x265-params,
  device tests use v4l2 on Linux instead of dshow
- Fix AAC test type assertions (uint16 → uint8 for Codec.Channels)
- Remove dead TestStripUserinfo referencing deleted function
- Fix creds test expectation to match current SecretString behavior
- Fix TLV8 round-trip tests: marshal normalizes separator 0x00 → 0xFF
- Fix hap/camera tests: renamed SetupEndpoints → SetupEndpointsResponse,
  mark TLV8 separator-affected round-trips as noequal
- Fix hap/hds test: Client() renamed to NewConn()
- Fix streams tests: register required scheme handlers, add cleanup

+++ pkg/aac/aac_test.go
@@ -17,7 +17,7 @@ func TestConfigToCodec(t *testing.T) {
 	codec := ConfigToCodec(src)
 	require.Equal(t, core.CodecAAC, codec.Name)
 	require.Equal(t, uint32(24000), codec.ClockRate)
-	require.Equal(t, uint16(1), codec.Channels)
+	require.Equal(t, uint8(1), codec.Channels)
 
 	dst := EncodeConfig(TypeAACELD, 24000, 1, true)
 	require.Equal(t, src, dst)
@@ -31,7 +31,7 @@ func TestADTS(t *testing.T) {
 
 	codec := ADTSToCodec(src)
 	require.Equal(t, uint32(44100), codec.ClockRate)
-	require.Equal(t, uint16(2), codec.Channels)
+	require.Equal(t, uint8(2), codec.Channels)
 
 	size := ReadADTSSize(src)
 	require.Equal(t, uint16(16), size)
